[DE-801] Fix invoice events (#35)
- Remove default collection method value
- Reorganize `Invoice Event` structure:
  - Invoice Event is now union type
  - Invoice Event Data is no longer union type. Each Invoice Event has it's own invoice_data with a concrete type
  - Added new event_types
- Removed some redundant classes
- Set usage ids, events ids and invoice event ids as longs,
- Set Invoice Event id to long
- Add handle parameter for create product family request body
- Add in_use query parameter for listing subscription components

+++ AdvancedBilling.Standard/Models/ListUsagesInput.cs
@@ -14,7 +14,6 @@ namespace AdvancedBilling.Standard.Models
     using AdvancedBilling.Standard;
     using AdvancedBilling.Standard.Models.Containers;
     using AdvancedBilling.Standard.Utilities;
-    using JsonSubTypes;
     using Newtonsoft.Json;
     using Newtonsoft.Json.Converters;
 
@@ -44,8 +43,8 @@ namespace AdvancedBilling.Standard.Models
         public ListUsagesInput(
             int subscriptionId,
             ListUsagesInputComponentId componentId,
-            int? sinceId = null,
-            int? maxId = null,
+            long? sinceId = null,
+            long? maxId = null,
             DateTime? sinceDate = null,
             DateTime? untilDate = null,
             int? page = 1,
@@ -77,13 +76,13 @@ namespace AdvancedBilling.Standard.Models
         /// Returns usages with an id greater than or equal to the one specified
         /// </summary>
         [JsonProperty("since_id", NullValueHandling = NullValueHandling.Ignore)]
-        public int? SinceId { get; set; }
+        public long? SinceId { get; set; }
 
         /// <summary>
         /// Returns usages with an id less than or equal to the one specified
         /// </summary>
         [JsonProperty("max_id", NullValueHandling = NullValueHandling.Ignore)]
-        public int? MaxId { get; set; }
+        public long? MaxId { get; set; }
 
         /// <summary>
         /// Returns usages with a created_at date greater than or equal to midnight (12:00 AM) on the date specified.
